+++ src/frontend/src/pages/ProjectDetailPage/ProjectForm/ProjectCreateContainer.tsx
@@ -2,6 +2,7 @@
  * This file is part of NER's FinishLine and licensed under GNU AGPLv3.
  * See the LICENSE file in the repository root folder for details.
  */
+import { wbsTester } from '../../../utils/form';
 import { useAllLinkTypes, useCreateSingleProject } from '../../../hooks/projects.hooks';
 import { useToast } from '../../../hooks/toasts.hooks';
 import { CreateSingleProjectPayload } from '../../../utils/types';
@@ -14,12 +15,18 @@ import { getRequiredLinkTypeNames } from '../../../utils/link.utils';
 import ErrorPage from '../../ErrorPage';
 import LoadingIndicator from '../../../components/LoadingIndicator';
 import { useQuery } from '../../../hooks/utils.hooks';
+import { useForm } from 'react-hook-form';
 import * as yup from 'yup';
+import { FormInput } from '../../CreateChangeRequestPage/CreateChangeRequestView';
+import { StandardChangeRequestType } from '../../CreateChangeRequestPage/CreateChangeRequestView';
 import { ProjectCreateChangeRequestFormInput } from './ProjectEditContainer';
 import { ProjectProposedChangesCreateArgs, WbsNumber, WorkPackageStage } from 'shared';
 import { CreateStandardChangeRequestPayload, useCreateStandardChangeRequest } from '../../../hooks/change-requests.hooks';
 import { useCreateSingleWorkPackage } from '../../../hooks/work-packages.hooks';
 import { useGetAllCars } from '../../../hooks/cars.hooks';
+import { ChangeRequestReason } from 'shared';
+import { yupResolver } from '@hookform/resolvers/yup';
+import { ChangeRequestType } from 'shared';
 
 const ProjectCreateContainer: React.FC = () => {
   const toast = useToast();
@@ -35,28 +42,6 @@ const ProjectCreateContainer: React.FC = () => {
   const { mutateAsync: mutateCRAsync, isLoading: isCRHookLoading } = useCreateStandardChangeRequest();
   const { mutateAsync: createWpMutateAsync, isLoading: createWpIsLoading } = useCreateSingleWorkPackage();
 
-  const {
-    data: allLinkTypes,
-    isLoading: allLinkTypesIsLoading,
-    isError: allLinkTypesIsError,
-    error: allLinkTypesError
-  } = useAllLinkTypes();
-
-  if (
-    createProjectIsLoading ||
-    isCRHookLoading ||
-    createWpIsLoading ||
-    !allLinkTypes ||
-    allLinkTypesIsLoading ||
-    carsIsLoading ||
-    !cars
-  )
-    return <LoadingIndicator />;
-  if (allLinkTypesIsError) return <ErrorPage message={allLinkTypesError.message} />;
-
-  if (carsIsError) return <ErrorPage message={carsError.message} />;
-
-  const requiredLinkTypeNames = getRequiredLinkTypeNames(allLinkTypes);
   const defaultValues = {
     name: '',
     budget: 0,
@@ -86,6 +71,78 @@ const ProjectCreateContainer: React.FC = () => {
     workPackages: yup.array()
   });
 
+  const changeRequestSchema = yup.object().shape({
+    type: yup.mixed<StandardChangeRequestType>().required('Type is required'),
+    what: yup.string().required('What is required'),
+    why: yup
+      .array()
+      .min(1, 'At least one Why is required')
+      .required('Why is required')
+      .of(
+        yup.object().shape({
+          type: yup.mixed<ChangeRequestReason>().required('Why Type is required'),
+          explain: yup
+            .string()
+            .required('Why Explain is required')
+            .when('type', ([type], schema) =>
+              type === ChangeRequestReason.OtherProject
+                ? schema.required().test('wbs-num-valid', 'WBS Number is not valid', wbsTester)
+                : yup.string()
+            )
+        })
+      )
+  });
+
+  const { reset: resetChangeRequestForm, ...changeRequestFormMethods } = useForm<FormInput>({
+    resolver: yupResolver(changeRequestSchema),
+    defaultValues: query.get('budgetChange')
+      ? {
+          what: 'Increase the budget to account for the cost of materials',
+          why: [{ type: ChangeRequestReason.Other, explain: 'The cost of materials ended up exceeding the initial budget' }],
+          type: ChangeRequestType.Issue
+        }
+      : query.get('timelineDelay')
+        ? {
+            what: 'Timeline delay',
+            why: [{ type: ChangeRequestReason.Other, explain: 'Decided to extend timeline after design review' }],
+            type: ChangeRequestType.Redefinition
+          }
+        : query.get('createWP')
+          ? {
+              what: '',
+              why: [{ type: ChangeRequestReason.Initialization, explain: 'Creating a Work Package on this Project' }],
+              type: ChangeRequestType.Redefinition
+            }
+          : {
+              what: '',
+              why: [{ type: ChangeRequestReason.Other, explain: '' }],
+              type: ChangeRequestType.Issue
+            }
+  });
+
+  const {
+    data: allLinkTypes,
+    isLoading: allLinkTypesIsLoading,
+    isError: allLinkTypesIsError,
+    error: allLinkTypesError
+  } = useAllLinkTypes();
+
+  if (
+    createProjectIsLoading ||
+    isCRHookLoading ||
+    createWpIsLoading ||
+    !allLinkTypes ||
+    allLinkTypesIsLoading ||
+    carsIsLoading ||
+    !cars
+  )
+    return <LoadingIndicator />;
+  if (allLinkTypesIsError) return <ErrorPage message={allLinkTypesError.message} />;
+
+  if (carsIsError) return <ErrorPage message={carsError.message} />;
+
+  const requiredLinkTypeNames = getRequiredLinkTypeNames(allLinkTypes);
+
   const onSubmitChangeRequest = async (data: ProjectCreateChangeRequestFormInput) => {
     const { name, budget, summary, links, teamIds, carNumber, descriptionBullets, type, what, why } = data;
 
@@ -96,7 +153,7 @@ const ProjectCreateContainer: React.FC = () => {
       const projectPayload: ProjectProposedChangesCreateArgs = {
         name,
         summary,
-        teamIds: teamIds.map((number) => '' + number),
+        teamIds: teamIds.map((id: string) => '' + id),
         budget,
         descriptionBullets,
         links,
@@ -224,6 +281,7 @@ const ProjectCreateContainer: React.FC = () => {
       managerId={managerId}
       onSubmitChangeRequest={onSubmitChangeRequest}
       setCarNumber={setCarNumber}
+      changeRequestFormReturn={changeRequestFormMethods}
     />
   );
 };
